Update Archimedean Spiral path implementation and tests

+++ src/fl/draw_visitor.cpp
@@ -9,7 +9,7 @@ namespace fl {
 XYDrawComposited::XYDrawComposited(const CRGB &color, const XYMap &xymap, CRGB *out)
     : mColor(color), mXYMap(xymap), mOut(out) {}
 
-void XYDrawComposited::draw(const point_xy<uint16_t> &pt, uint32_t index, uint8_t value) {
+void XYDrawComposited::draw(const point_xy<int> &pt, uint32_t index, uint8_t value) {
     FASTLED_UNUSED(pt);
     CRGB& c = mOut[index];
     CRGB blended = mColor;
