+++ src/fl/draw_visitor.h
@@ -12,14 +12,14 @@ namespace fl {
 // A visitor interface for drawing uint8_t values.
 struct XYDrawUint8Visitor {
     virtual ~XYDrawUint8Visitor() = default;
-    virtual void draw(const point_xy<uint16_t> &pt, uint32_t index, uint8_t value) = 0;
+    virtual void draw(const point_xy<int> &pt, uint32_t index, uint8_t value) = 0;
 };
 
 
 // Draws a uint8_t value to a CRGB array, blending it with the existing color.
 struct XYDrawComposited: public XYDrawUint8Visitor {
     XYDrawComposited(const CRGB &color, const XYMap &xymap, CRGB *out);
-    void draw(const point_xy<uint16_t> &pt, uint32_t index, uint8_t value) override ;
+    void draw(const point_xy<int> &pt, uint32_t index, uint8_t value) override ;
     const CRGB mColor;
     const XYMap mXYMap;
     CRGB *mOut;
