+++ src/fl/raster.h
@@ -21,11 +21,11 @@ class SubPixel2x2;
 class Raster {
   public:
     Raster() = default;
-    Raster(const point_xy<uint16_t> &origin, uint16_t width, uint16_t height) {
+    Raster(const point_xy<int> &origin, uint16_t width, uint16_t height) {
         reset(origin, width, height);
     }
     Raster(const Raster &) = delete;
-    void reset(const point_xy<uint16_t> &origin, uint16_t width,
+    void reset(const point_xy<int> &origin, uint16_t width,
                uint16_t height) {
         mGrid.reset(width, height);
         mOrigin = origin;
@@ -38,16 +38,16 @@ class Raster {
     uint8_t &at(uint16_t x, uint16_t y) { return mGrid.at(x, y); }
     const uint8_t &at(uint16_t x, uint16_t y) const { return mGrid.at(x, y); }
 
-    const point_xy<uint16_t> origin() const { return mOrigin; }
-    point_xy<uint16_t> global_min() const { return mOrigin; }
-    point_xy<uint16_t> global_max() const {
-        return mOrigin + point_xy<uint16_t>(mGrid.width(), mGrid.height());
+    const point_xy<int> origin() const { return mOrigin; }
+    point_xy<int> global_min() const { return mOrigin; }
+    point_xy<int> global_max() const {
+        return mOrigin + point_xy<int>(mGrid.width(), mGrid.height());
     }
 
-    rect_xy<uint16_t> bounds() const {
-        point_xy<uint16_t> min = origin();
-        point_xy<uint16_t> max = global_max();
-        return rect_xy<uint16_t>(min, max);
+    rect_xy<int> bounds() const {
+        point_xy<int> min = origin();
+        point_xy<int> max = global_max();
+        return rect_xy<int>(min, max);
     }
 
     uint16_t width() const { return mGrid.width(); }
@@ -62,7 +62,7 @@ class Raster {
 
   private:
     Grid<uint8_t> mGrid;
-    point_xy<uint16_t> mOrigin;
+    point_xy<int> mOrigin;
 };
 
 } // namespace fl
